refactor: move API URL logic to plugin
